# public/openapi.yaml
openapi: 3.0.3
info:
  title: CRSET Solutions API
  description: |
    API completa para integracao com a plataforma CRSET Solutions.
    
    Inclui endpoints para:
    - Chat AGI (OpenAI GPT-4o-mini)
    - RAG (Retrieval-Augmented Generation)
    - Formulario de contacto
    - Autenticacao e status
    
    **Base URL:** https://crsetsolutions.com
  version: 1.0.0
  contact:
    name: CRSET Solutions
    email: crsetsolutions@gmail.com
    url: https://crsetsolutions.com
  license:
    name: Proprietary
    url: https://crsetsolutions.com/termos
servers:
  - url: https://crsetsolutions.com
    description: Producao
  - url: http://localhost:3000
    description: Desenvolvimento local
...
paths:
...
  /api/agi/chat:
    post:
      summary: Chat com AGI
      description: Envia uma mensagem para o modelo GPT-4o-mini e recebe uma resposta.
      tags:
        - AGI
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Mensagem do utilizador
                  example: "Ola, como posso criar um site profissional?"
                strict:
                  type: boolean
                  description: Desativar persona (modo raw)
                  default: false
                mode:
                  type: string
                  enum: [normal, raw]
                  description: Modo de resposta
                  default: normal
      responses:
        '200':
          description: Resposta gerada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  reply:
                    type: string
                    example: "Ola! Para criar um site profissional..."
                  meta:
                    type: object
                    properties:
                      backend:
                        type: string
                        example: openai
                      upstream_base:
                        type: string
                        example: https://api.openai.com/v1
                      model:
                        type: string
                        example: gpt-4o-mini
                      agent:
                        type: string
                        example: boris
        '403':
          description: Autenticacao necessaria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Requisicao invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/contact:
    post:
      summary: Enviar formulario de contacto
      description: Envia um lead atraves do formulario de contacto.
      tags:
        - Contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - message
              properties:
                name:
                  type: string
                  description: Nome do contacto
                  example: "Joao Silva"
                email:
                  type: string
                  format: email
                  description: Email do contacto
                  example: "joao.silva@example.com"
                message:
                  type: string
                  description: Mensagem
                  example: "Gostaria de saber mais sobre os servicos."
                utm:
                  type: object
                  description: Parametros UTM para tracking
                  properties:
                    source:
                      type: string
                      example: "google"
                    medium:
                      type: string
                      example: "cpc"
                    campaign:
                      type: string
                      example: "servicos"
                    content:
                      type: string
                      example: "essential"
      responses:
        '200':
          description: Mensagem enviada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
        '400':
          description: Dados invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
          example: "bad_request"
        message:
          type: string
          example: "Missing required field: message"
        reason:
          type: string
          description: Razao especifica do erro (opcional)
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: chat_flag
      description: Cookie de autenticacao JWT para acesso ao chat AGI